+++ src/components/SearchSong/index.tsx
@@ -25,18 +25,7 @@ const SearchSong = ({ item }: Props) => {
           />
           {}
           <button className="w-12 h-12 bg-primary right-2 bottom-2 transition-opacity flex items-center shadow-2xl justify-center opacity-0 opacity-100 ">
-            <svg
-              role="img"
-              height="24"
-              width="24"
-              viewBox="0 0 24 24"
-              className="Svg-sc-1bi12j5-0 hDgDGI"
-            >
-              <path
-                d="M7.05 3.606l13.49 7.788a.7.7 0 010 1.212L7.05 20.394A.7.7 0 016 19.788V4.212a.7.7 0 011.05-.606z"
-                fill="black"
-              ></path>
-            </svg>
+           
           </button>
         </div>
         <div className="flex-5 ">
