+++ src/components/createPlaylistSong/index.tsx
@@ -1,8 +1,8 @@
-
 import { useDispatch } from "react-redux";
 import { setRenderState } from "../../redux/slices/auth";
 import { AddItemToPL } from "../../services/addItemsToPlaylist";
 import { getPlaylist } from "../../services/getPlaylist";
+import ButtonItem from "../buttonItem";
 import "./index.css";
 type Props = { item: any };
 
@@ -11,7 +11,7 @@ const CreatePlaylistSong = ({ item }: Props) => {
   const addItemToPlayList = (itemURI: string) => {
     dispatch(setRenderState());
     AddItemToPL(itemURI);
-    getPlaylist()
+    getPlaylist();
   };
   return (
     //
@@ -36,21 +36,11 @@ const CreatePlaylistSong = ({ item }: Props) => {
             height="40px"
             width="40px"
           />
-          {}
-          <button className="w-12 h-12 bg-primary right-2 bottom-2 transition-opacity flex items-center shadow-2xl justify-center opacity-0 opacity-100 ">
-            <svg
-              role="img"
-              height="24"
-              width="24"
-              viewBox="0 0 24 24"
-              className="Svg-sc-1bi12j5-0 hDgDGI"
-            >
-              <path
-                d="M7.05 3.606l13.49 7.788a.7.7 0 010 1.212L7.05 20.394A.7.7 0 016 19.788V4.212a.7.7 0 011.05-.606z"
-                fill="black"
-              ></path>
-            </svg>
+          <button className="w-12 h-12 bg-primary right-2 bottom-2 transition-opacity flex items-center shadow-2xl justify-center  opacity-0 opacity-100 ">
+            <ButtonItem />
           </button>
+
+          {/* button */}
         </div>
         <div className="flex-5 ">
           <span className=" font-semibold text-white text-[16px] whitespace-nowrap flex">
